update to rust 1.16.0

+++ rust/modulino.rs
@@ -1,9 +1,7 @@
 //! Modulino
 
-#![crate_id(name = "modulino")]
-
 // A dummy function to import into test.rs
-pub fn meaning_of_life() -> int {
+pub fn meaning_of_life() -> i64 {
   //! The answer to life, the universe, and everything.
   return 42;
 }
